Add compile-commands watcher
The compile commands watcher will invoke build-compile-commands whenever
a files is added/removed from the src/include directories or when a
Makefile is modified (for PROS projects). The goal is to maintain an
up-to-date compile-commands.json file. Eventually, I'd like to move all
of this file-watching logic into the CLI, but I'm having trouble finding
a Python package to do so.