Compile commands watcher (#11)
### Summary

The compile commands watcher will invoke build-compile-commands whenever
a files is added/removed from the src/include directories or when a
Makefile is modified (for PROS projects). The goal is to maintain an
up-to-date compile-commands.json file. Eventually, I'd like to move all
of this file-watching logic into the CLI, but I'm having trouble finding
a Python package to do so.

The most interesting thing of note in this PR is that I'm needing to "debounce" when we build
the compile commands, so that quickly adding (or copying) multiple files into a project doesn't
cause build-compile-commands to be invoked twice.

Also, removed some debugging print statements

### Test Plan

- [X] Creating a project doesn't invoke this facility
- [X] Adding a file to an existing project invokes build-compile-commands (using Atom menus)
- [X] Adding a file to a newly created project invokes build-compile-commands (using Atom menus)
- [X] Adding several files doesn't invoke build-compile-commands multiple times.

#### References

T527, T572